+++ packages/core/src/utils/create-registry.ts
@@ -1,19 +1,11 @@
-import {
-  type IMessageTypeRegistry,
-  createDescriptorSet,
-  createRegistryFromDescriptors,
-} from "@bufbuild/protobuf";
-import { expect, test } from "vitest";
+import { type IMessageTypeRegistry, createDescriptorSet, createRegistryFromDescriptors } from "@bufbuild/protobuf";
 
-import { FileDescriptorProto } from "@bufbuild/protobuf";
+import type { FileDescriptorProto } from "@bufbuild/protobuf";
 import type { Package } from "../proto.js";
 
 export function createRegistry(substream: Package): IMessageTypeRegistry {
   substream.protoFiles = topoSort(substream.protoFiles);
-  return createRegistryFromDescriptors(
-    createDescriptorSet(substream.protoFiles),
-    true
-  );
+  return createRegistryFromDescriptors(createDescriptorSet(substream.protoFiles), true);
 }
 
 // createDescriptorSet expects the files to be topologically sorted
@@ -23,13 +15,15 @@ function topoSort(protoFiles: FileDescriptorProto[]): FileDescriptorProto[] {
 
   // Build dependency graph
   for (const file of protoFiles) {
-    if (!file.name) continue;
+    if (!file.name) {
+      continue;
+    }
     graph.set(file.name, new Set(file.dependency));
     for (const dep of file.dependency) {
       if (!deps.has(dep)) {
         deps.set(dep, new Set());
       }
-      deps.get(dep)!.add(file.name);
+      deps.get(dep)?.add(file.name);
     }
   }
 
@@ -47,11 +41,20 @@ function topoSort(protoFiles: FileDescriptorProto[]): FileDescriptorProto[] {
 
     // Check what nodes are now ready
     if (deps.has(current.name)) {
-      for (const dependent of deps.get(current.name)!) {
+      const current_deps = deps.get(current.name);
+      if (!current_deps) {
+        throw new Error(`Missing dependency for ${current.name}`);
+      }
+      for (const dependent of current_deps) {
         const depNode = [...remaining].find((f) => f.name === dependent);
-        if (!depNode) continue;
+        if (!depNode) {
+          continue;
+        }
 
-        const depGraph = graph.get(dependent)!;
+        const depGraph = graph.get(dependent);
+        if (!depGraph) {
+          throw new Error(`Missing graph for ${dependent}`);
+        }
         depGraph.delete(current.name);
 
         if (depGraph.size === 0) {
@@ -64,26 +67,6 @@ function topoSort(protoFiles: FileDescriptorProto[]): FileDescriptorProto[] {
   return ordered;
 }
 
-const mockProtoFile1 = new FileDescriptorProto({
-  name: "file1.proto",
-  dependency: [],
-});
-
-const mockProtoFile2 = new FileDescriptorProto({
-  name: "file2.proto",
-  dependency: ["file1.proto"],
-});
-
-const mockProtoFile3 = new FileDescriptorProto({
-  name: "file3.proto",
-  dependency: ["file2.proto", "file1.proto"],
-});
-
-test("topoSort orders files by dependencies", () => {
-  const files = [mockProtoFile3, mockProtoFile1, mockProtoFile2];
-  const sorted = topoSort(files);
-
-  expect(sorted[0]).toBe(mockProtoFile1);
-  expect(sorted[1]).toBe(mockProtoFile2);
-  expect(sorted[2]).toBe(mockProtoFile3);
-});
+if (process.env.NODE_ENV === "test") {
+  module.exports.topoSort = topoSort;
+}
